Refactored ccode by introducing base class CodePrinter
Methods related to analyzation of array expressions have been moved up
and adapted.  All tests pass.
